% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering_functions.R
\name{magnemite_adjust_actual_times}
\alias{magnemite_adjust_actual_times}
\title{Adjust Assigned Actual Times}
\usage{
magnemite_adjust_actual_times(
  tt,
  trace = c("top", "bot", "both"),
  direction = c("<-", "->"),
  amount = 1
)
}
\arguments{
\item{tt}{Timing tick object.}

\item{trace}{Which trace(s) to adjust: `"top"`, `"bot"`, or `"both"`.}

\item{direction}{Shift direction: `"<-"` or `"->"`.}

\item{amount}{Integer number of hours to shift.}
}
\value{
Updated timing tick object.
}
\description{
Shifts assigned `actual_time` values for top, bottom, or both traces by a
specified direction and amount.
}
