% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_clustering_functions.R
\name{magnemite_assign_actual_times}
\alias{magnemite_assign_actual_times}
\title{Assign Actual Times to Timing Ticks}
\usage{
magnemite_assign_actual_times(tt, min_length = 25)
}
\arguments{
\item{tt}{Timing tick object with top/bottom tables.}

\item{min_length}{Minimum sequence length used for time generation.}
}
\value{
Updated timing tick object.
}
\description{
Adds hourly `actual_time` labels to top and/or bottom timing tick tables.
}
