% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clippng_app.R
\name{magnemite_default_clippng_project_dir}
\alias{magnemite_default_clippng_project_dir}
\title{Resolve Default Clippng Project Directory}
\usage{
magnemite_default_clippng_project_dir()
}
\value{
Normalized project directory path.
}
\description{
Uses bundled database location when available, otherwise returns a fallback
project directory.
}
