% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clippng_app.R
\name{magnemite_find_package_file}
\alias{magnemite_find_package_file}
\title{Find a Bundled Package File}
\usage{
magnemite_find_package_file(...)
}
\arguments{
\item{...}{Path components relative to package root.}
}
\value{
Normalized file path, or `NA_character_` when not found.
}
\description{
Searches installed and development package locations for a relative file path.
}
