% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing_ticks_paths.R
\name{magnemite_list_tif_files}
\alias{magnemite_list_tif_files}
\title{List TIF Files}
\usage{
magnemite_list_tif_files(server_dir)
}
\arguments{
\item{server_dir}{Directory containing source image files.}
}
\value{
Character vector of `.tif` file paths.
}
\description{
Lists `.tif` files from a server directory.
}
