% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing_ticks_clustering.R
\name{magnemite_process_top_trace_cluster}
\alias{magnemite_process_top_trace_cluster}
\title{Process Top Trace Cluster}
\usage{
magnemite_process_top_trace_cluster(tif_path, rds_path)
}
\arguments{
\item{tif_path}{Path to the source `.tif` image.}

\item{rds_path}{Path to the digitized trace `.rds` file.}
}
\value{
A list with clustered top-trace data components.
}
\description{
Convenience wrapper to process the top trace region.
}
