% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing_ticks_apps.R
\name{magnemite_timing_tick_click_app}
\alias{magnemite_timing_tick_click_app}
\title{Build Timing Tick Click Correction App}
\usage{
magnemite_timing_tick_click_app(server_dir = NULL, timing_ticks_dir = NULL)
}
\arguments{
\item{server_dir}{Optional server source directory override.}

\item{timing_ticks_dir}{Optional timing tick RDS directory override.}
}
\value{
A Shiny app object.
}
\description{
Constructs the Shiny app used to manually inspect and adjust timing tick
positions.
}
