% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{magnemite_trace_medians_by_x}
\alias{magnemite_trace_medians_by_x}
\title{Compute Median Trace Y by X}
\usage{
magnemite_trace_medians_by_x(trace_df)
}
\arguments{
\item{trace_df}{Data frame containing at least `X` and `Y` columns.}
}
\value{
A data frame with columns `X` and `median_Y`.
}
\description{
Summarizes a trace data frame to median y values per x coordinate.
}
