% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clippng_app.R
\name{run_magnemite_clippng_app}
\alias{run_magnemite_clippng_app}
\title{Run Clippng Shiny App}
\usage{
run_magnemite_clippng_app(
  project_dir = NULL,
  db_path = NULL,
  output_dir = NULL
)
}
\arguments{
\item{project_dir}{Optional project directory override.}

\item{db_path}{Optional database path override.}

\item{output_dir}{Optional output directory override.}
}
\value{
The result of `shiny::runApp()`.
}
\description{
Launches the clipping app.
}
